Begin refactor of AbstractDBC (does not compile)
