+++ SpellGUIV2/Sources/DBC/AbstractDBC.cs
@@ -80,278 +80,12 @@ namespace SpellEditor.Sources.DBC
 
         public Task ImportTo(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc UpdateProgress, string IdKey, string bindingName, ImportExportType _type)
         {
-            switch (_type)
-            {
-                case ImportExportType.DBC:
-                    return ImportToSql(adapter, UpdateProgress, IdKey, bindingName);
-                case ImportExportType.CSV:
-                    return ImportToCsv(adapter, UpdateProgress, IdKey, bindingName);
-                // TODO: Handle import of raw SQL
-                default:
-                    throw new Exception("Unhandled ImportExport type: " + _type.ToString());
-            }
-        }
-
-        private Task ImportToCsv(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc UpdateProgress, string IdKey, string bindingName)
-        {
-            return Task.Run(() =>
-            {
-                var binding = BindingManager.GetInstance().FindBinding(bindingName);
-                adapter.Execute(string.Format(adapter.GetTableCreateString(binding), binding.Name.ToLower()));
-
-                var path = $"{Config.Config.DbcDirectory}\\{binding.Name}.csv";
-                if (!File.Exists(path))
-                    throw new Exception("Cannot find file: " + path);
-
-                var records = File.ReadAllLines(path);
-                records = records.Skip(1).ToArray(); // skip header
-
-                int count = records.Length;
-                int updateRate = count < 100 ? 100 : count / 100;
-                int index = 0;
-                StringBuilder q = null;
-                foreach (var record in records)
-                {
-                    if (index == 0 || index % 250 == 0)
-                    {
-                        if (q != null)
-                        {
-                            q.Remove(q.Length - 2, 2);
-                            adapter.Execute(q.ToString());
-                        }
-                        q = new StringBuilder();
-                        q.Append(string.Format("INSERT INTO `{0}` VALUES ", bindingName.ToLower()));
-                    }
-                    if (++index % updateRate == 0)
-                    {
-                        // Visual studio says these casts are redundant but it does not work without them
-                        double percent = (double)index / (double)count;
-                        UpdateProgress(percent);
-                    }
-                    q.Append("(");
-                    q.Append(record.Replace("\\r", "\r").Replace("\\n", "\n"));
-                    q.Append("), ");
-                }
-                if (q.Length > 0)
-                {
-                    q.Remove(q.Length - 2, 2);
-                    adapter.Execute(q.ToString());
-                }
-            });
-        }
-
-        private Task ImportToSql(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc UpdateProgress, string IdKey, string bindingName)
-        {
-            return Task.Run(() =>
-            {
-                var binding = BindingManager.GetInstance().FindBinding(bindingName);
-
-                adapter.Execute(string.Format(adapter.GetTableCreateString(binding), binding.Name.ToLower()));
-                uint currentRecord = 0;
-                uint count = Header.RecordCount;
-                uint updateRate = count < 100 ? 100 : count / 100;
-                uint index = 0;
-                StringBuilder q = null;
-                foreach (var recordMap in Body.RecordMaps)
-                {
-                    if (index == 0 || index % 250 == 0)
-                    {
-                        if (q != null)
-                        {
-                            q.Remove(q.Length - 2, 2);
-                            adapter.Execute(q.ToString());
-                        }
-                        q = new StringBuilder();
-                        q.Append(string.Format("INSERT INTO `{0}` VALUES ", bindingName.ToLower()));
-                    }
-                    if (++index % updateRate == 0)
-                    {
-                        // Visual studio says these casts are redundant but it does not work without them
-                        double percent = (double)index / (double)count;
-                        UpdateProgress(percent);
-                    }
-                    currentRecord = recordMap.ContainsKey(IdKey) ? (uint)recordMap[IdKey] : 0;
-                    q.Append("(");
-                    foreach (var field in binding.Fields)
-                    {
-                        switch (field.Type)
-                        {
-                            case BindingType.INT:
-                            case BindingType.UINT:
-                            case BindingType.UINT8:
-                                {
-                                    q.Append(string.Format("'{0}', ", recordMap[field.Name]));
-                                    break;
-                                }
-                            case BindingType.FLOAT:
-                            case BindingType.DOUBLE:
-                                {
-                                    q.Append(string.Format("REPLACE('{0}', ',', '.'), ", recordMap[field.Name]));
-                                    break;
-                                }
-                            case BindingType.STRING_OFFSET:
-                                {
-                                    var strOffset = (uint)recordMap[field.Name];
-                                    var lookupResult = Reader.LookupStringOffset(strOffset);
-                                    q.Append(string.Format("'{0}', ", adapter.EscapeString(lookupResult)));
-                                    break;
-                                }
-                            case BindingType.UNKNOWN:
-                                break;
-                            default:
-                                throw new Exception($"ERROR: Record[{currentRecord}] Unhandled type: {field.Type} on field: {field.Name}");
-                        }
-                    }
-                    q.Remove(q.Length - 2, 2);
-                    q.Append("), ");
-                }
-                if (q.Length > 0)
-                {
-                    q.Remove(q.Length - 2, 2);
-                    adapter.Execute(q.ToString());
-                }
-                // We have attempted to import the Spell.dbc so clean up unneeded data
-                // This will be recreated if the import process is started again
-                Reader.CleanStringsMap();
-            });
+            return StorageFactory.Instance.GetStorageAdapter(_type).Import(adapter, UpdateProgress, IdKey, bindingName);
         }
 
         public Task ExportTo(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc updateProgress, string IdKey, string bindingName, ImportExportType _type)
         {
-            switch (_type)
-            {
-                case ImportExportType.DBC:
-                    return ExportToDbc(adapter, updateProgress, IdKey, bindingName);
-                case ImportExportType.CSV:
-                    return ExportToCsv(adapter, updateProgress, IdKey, bindingName);
-                case ImportExportType.SQL:
-                    return ExportToSql(adapter, updateProgress, bindingName);
-                default:
-                    throw new Exception("Unhandled ImportExport type: " + _type.ToString());
-            }
-        }
-
-        public Task ExportToDbc(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc updateProgress, string IdKey, string bindingName)
-        {
-            return Task.Run(() =>
-            {
-                var binding = BindingManager.GetInstance().FindBinding(bindingName);
-                if (binding == null)
-                    throw new Exception("Binding not found: " + bindingName);
-                var body = new DBCBodyToSerialize();
-
-                var orderClause = "";
-                if (binding.OrderOutput)
-                {
-                    orderClause = binding.Fields.FirstOrDefault(f => f.Name.Equals(IdKey)) != null ? $" ORDER BY `{IdKey}`" : "";
-                }
-
-                body.Records = LoadRecords(adapter, bindingName, orderClause, updateProgress);
-                var numRows = body.Records.Count();
-                if (numRows == 0)
-                    throw new Exception("No rows to export");
-
-                Header = new DBCHeader
-                {
-                    FieldCount = (uint)binding.Fields.Count(),
-                    // Magic is always 'WDBC' https://wowdev.wiki/DBC
-                    Magic = 1128416343,
-                    RecordCount = (uint)numRows,
-                    RecordSize = (uint)binding.CalcRecordSize(),
-                    StringBlockSize = body.GenerateStringOffsetsMap(binding)
-                };
-
-                SaveDbcFile(updateProgress, body, binding);
-            });
-        }
-
-        public Task ExportToSql(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc updateProgress, string bindingName)
-        {
-            return Task.Run(() =>
-            {
-                using (var newAdapter = AdapterFactory.Instance.GetAdapter(false))
-                {
-                    adapter.ExportTableToSql(bindingName, "Export", Task.CurrentId, updateProgress);
-                }
-            });
-        }
-
-        public Task ExportToCsv(IDatabaseAdapter adapter, MainWindow.UpdateProgressFunc updateProgress, string IdKey, string bindingName)
-        {
-            return Task.Run(() =>
-            {
-                var binding = BindingManager.GetInstance().FindBinding(bindingName);
-                if (binding == null)
-                    throw new Exception("Binding not found: " + bindingName);
-                var body = new DBCBodyToSerialize();
-
-                var orderClause = "";
-                if (binding.OrderOutput)
-                {
-                    orderClause = binding.Fields.FirstOrDefault(f => f.Name.Equals(IdKey)) != null ? $" ORDER BY `{IdKey}`" : "";
-                }
-
-                body.Records = LoadRecords(adapter, bindingName, orderClause, updateProgress);
-                var numRows = body.Records.Count();
-                if (numRows == 0)
-                    throw new Exception("No rows to export");
-
-                string path = $"Export/{binding.Name}.csv";
-                Directory.CreateDirectory(Path.GetDirectoryName(path));
-                if (File.Exists(path))
-                    File.Delete(path);
-
-                var toWrite = new List<string>(body.Records.Count + 1)
-                {
-                    // Write header
-                    string.Join(",", binding.Fields.Select(f => f.Name).Select(f => EscapeCsv(f)).ToList())
-                };
-                body.Records.ForEach(record =>
-                {
-                    List<string> fields = new List<string>(binding.Fields.Length);
-                    foreach (var field in binding.Fields)
-                    {
-                        switch (field.Type)
-                        {
-                            case BindingType.STRING_OFFSET:
-                                {
-                                    var str = EscapeCsv(record[field.Name].ToString());
-                                    if (str.Length == 0)
-                                        str = "\"\"";
-                                    fields.Add(str);
-                                    break;
-                                }
-                            default:
-                                fields.Add(record[field.Name].ToString());
-                                break;
-
-                        }
-                    }
-                    toWrite.Add(string.Join(",", fields));
-                });
-
-                File.WriteAllText(path, string.Join("\n", toWrite));
-            });
-        }
-
-        private string EscapeCsv(string field)
-        {
-            var returnVal = field;
-            if (Regex.Matches(field, "([A-Z]|[a-z]|\n|\\\"|\'| |&|~|!|£|$|%|\\^|&|\\*|\\(|\\)|`|$)+").Count > 0)
-            {
-                StringBuilder sb = new StringBuilder();
-                sb.Append("\"");
-                foreach (char nextChar in field)
-                {
-                    sb.Append(nextChar);
-                    if (nextChar == '"')
-                        sb.Append("\"");
-                }
-                sb.Append("\"");
-                returnVal = sb.ToString();
-            }
-            returnVal = returnVal.Replace("\n", "\\n").Replace("\r", "\\r");
-            return returnVal;
+            return StorageFactory.Instance.GetStorageAdapter(_type).Export(adapter, updateProgress, IdKey, bindingName);
         }
 
         protected List<Dictionary<string, object>> LoadRecords(IDatabaseAdapter adapter, string bindingName, string orderClause, MainWindow.UpdateProgressFunc updateProgress)
