+++ SpellGUIV2/SpellEditor.csproj
@@ -212,6 +212,11 @@
     <Compile Include="Sources\Controls\Visual\VisualPasteListEntry.cs" />
     <Compile Include="Sources\Controls\Visual\VisualEffectListEntry.cs" />
     <Compile Include="Sources\Controls\Visual\VisualKitListEntry.cs" />
+    <Compile Include="Sources\Database\SqlFileStorage.cs" />
+    <Compile Include="Sources\Database\CsvFileStorage.cs" />
+    <Compile Include="Sources\Database\DbcFileStorage.cs" />
+    <Compile Include="Sources\Database\StorageAdapter.cs" />
+    <Compile Include="Sources\Database\StorageFactory.cs" />
     <Compile Include="Sources\Database\ImportExportType.cs" />
     <Compile Include="Sources\Database\AdapterFactory.cs" />
     <Compile Include="Sources\DBC\AnimationData.cs" />
